FreeStream.Core

Documentation

type Task a b m r = FreeT (TaskF a b) m r

A Task is the free monad transformer arising from TaskF.

newtype TaskF a b k

TaskF is the union of unary functions and binary products into a single type, here defined with a Boehm-Berarducci encoding. The value constructor functions are awaitF and yieldF.

Rather than using a normal ADT, which would certainly make the code a bit easier to read and write, a value of this type is actually a control flow mechanism accepting two continuations and choosing one or the other.

Client code should never actually have to deal with this.

Constructors

TaskF

Fields

runT :: forall r. ((a -> k) -> r) -> ((b, k) -> r) -> r

type Source b m r = forall x. Task x b m r

A computation which only yields and never awaits

type Sink a m r = forall x. Task a x m r

A computation which only awaits and never yields

type Action m r = forall x. Task x x m r

A computation which neither yields nor awaits

run :: FreeT f m a -> m (FreeF f a (FreeT f m a))

run is shorter than runFreeT and who knows, maybe it'll change some day

await :: Monad m => Task a b m a

Command to wait for a new value upstream

yield :: Monad m => b -> Task a b m ()

Command to send a value downstream

liftT :: (MonadTrans t, Monad m) => FreeT f m a -> t m (FreeF f a (FreeT f m a))

Simple utilities that make writing this library easier

each :: (Monad m, Foldable t) => t b -> Task a b m ()

Convert a list to a Source

for :: Monad m => Task a b m r -> (b -> Task a c m s) -> Task a c m r

Enumerate yielded values into a continuation, creating a new Source

(><) :: Monad m => Task a b m r -> Task b c m r -> Task a c m r infixl 3

Compose two tasks in a pull-based stream

(>-) :: Monad m => Task a b m r -> (b -> Task b c m r) -> Task a c m r

Connect a task to a continuation yielding another task; see ><

(~>) :: Monad m => Task a b m r -> (b -> Task a c m s) -> Task a c m r

Infix version of for