FreeStream.Par

Documentation

broadcast :: (Functor f, Monad m) => Task a b m r -> f (Task b c m r) -> f (Task a c m r)

Broadcast the values of a source Task to Functor of sink Tasks

merge :: (Foldable t, Monad m) => t (Task a c m s) -> Task a c m ()

Naively merges the values of some Functor of source Tasks

(*<) :: (Functor f, Monad m) => Task a b m r -> f (Task b c m r) -> f (Task a c m r) infixl 9

Infix synonym for broadcast

(>*) :: (Foldable t, Monad m) => t (Task a b m s) -> Task b c m () -> Task a c m () infixr 0

merges a Functor of source Tasks and pipes (><) to a sink Task